requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
pinecone-client==2.2.2
pinecone-text==0.5.2
sentence-transformers==2.2.2
torch==2.0.1
jsonpickle==3.0.1
python-dotenv==1.0.0
datasets==2.14.4
python-multipart==0.0.6
